requirements.txt
discord.py
asyncio
flask
